displayStringTableForSection(): add $content parameter.
This parameter allows the caller to specify the first words of the sentences
displayed by this function. If this extra parameter is omitted, the function
will default to "Strings". The point of this parameter is that ELF string
tables can hold various things that are not symbol names, e.g. file names or
versions.